requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.9
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
narwhals==1.38.2
packaging==25.0
plotly==6.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pyttsx3==2.98
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
google-cloud-texttospeech==2.16.3
google-api-core>=2.0.0
protobuf>=3.19.0
python-dotenv==1.0.1